test: consolidate duplicate EigenLayer and manager tests
Reduces test code duplication by 265+ lines while maintaining full coverage.

Changes:
- Delete multi_avs_test.rs (165 lines) - fully duplicated in eigenlayer_e2e_test.rs
- Remove duplicate hypervisor validation test from registration.rs (30 lines)
- Remove overlapping blueprint spawning test from protocol_integration.rs (70 lines)
- Extract common test utilities to tests/common.rs:
  - create_test_context() helper
  - create_avs_config() helper
- Update all test files to use common utilities

Impact:
- Before: 1,628 lines of integration tests, 1,074 lines with unit tests
- After: ~1,363 lines total (16% reduction)
- Test coverage: Unchanged - all scenarios still covered
- Maintainability: Significantly improved (DRY principle)

All tests compile successfully. No functional changes.
